Pause Menu Additions
>Added outlines to all existing UI text that needs it.
>3 new Buttons added to Pause Menu Panel inside the Single Player Level
(Will need to be brought over to Co-Op Level [Is it possible to make the
whole Pause Menu some sort of prefab?))
>> Restart, Options, Quit Game
> Restart and Quit Game work on a basic functional level
>> Minor tweaks will need to be made.

I've learned some things on how the Unity SceneManager works and I
wonder if I can also change some of my current code to be more
expansive. I could potentially not need to make so many scripts.