Working on UI
Got the Startmenu finished as previously stated.
Currently working on making the Pause Menu a prefab. I may have to
change how I want the pause menu to work possibly if I don't want to
change it a lot in it's current state but I've already added the Options
panel.
The only thing will be to create a modal button for Quit Game asking if
you want to quit the application or if you want to return to the main
menu. ...Also making the Options menu work as a real options menu.